It works?

+++ emulator/m64py/include/m64py_ssb64_memory.h
@@ -141,7 +141,7 @@ void
 m64py_memory_read_fighter_acceleration(m64py_memory_interface_t* memory, uint32_t fighter_address, float* xacc, float* yacc);
 
 void
-m64py_memory_read_fighter_orientation(m64py_memory_interface_t* memory, uint32_t fighter_address, int* orientation);
+m64py_memory_read_fighter_orientation(m64py_memory_interface_t* memory, uint32_t fighter_address, int32_t* orientation);
 
 void
 m64py_memory_read_fighter_movement_frame(m64py_memory_interface_t* memory, uint32_t fighter_address, uint32_t* frame);
@@ -156,7 +156,7 @@ void
 m64py_memory_read_fighter_shield_break_recovery_timer(m64py_memory_interface_t* memory, uint32_t fighter_address, uint32_t* time_left);
 
 void
-m64py_memory_read_fighter_percent(m64py_memory_interface_t* memory, uint32_t fighter_address, uint16_t* percent);
+m64py_memory_read_fighter_percent(m64py_memory_interface_t* memory, uint32_t fighter_address, uint32_t* percent);
 
 void
 m64py_memory_read_fighter_is_invincible(m64py_memory_interface_t* memory, uint32_t fighter_address, int* is_invincible);
